requirements.txt
Pillow>=10.0.0
PyQt5>=5.15.0
pyinstaller>=6.0.0
